backend/requirements.txt
beautifulsoup4==4.14.3
celery==5.6.2
fastapi==0.135.1
httpx==0.28.1
openai==2.26.0
pandas==3.0.1
playwright==1.58.0
protobuf==7.34.0
pydantic==2.12.5
PyJWT==2.11.0
python-dotenv==1.2.2
Requests==2.32.5
scikit_learn==1.8.0
sentence_transformers==5.2.3
SQLAlchemy==2.0.48
uvicorn==0.41.0
python-multipart
pg8000==1.31.2
openpyxl==3.1.5
